Cleaned up executeTurn.
Modified setupTurn so the CPU's have a delay on attacks when both players are CPUs, allowing the user to watch them battle.

Changed the gameOver function so it always takes playerOne and playerTwo as the first and second arguments in that order, and a third argument, playerOneWins, is a boolean that tells which player won. This is kind of an ugly band-aid but I prefer it over the alternatives with my current code