Tweaked the bot timing so I can watch the way they randomly shoot.
Refactored executeTurn because I changed the way Player.randomAttack works. Now, randomAttack calls sendAttack internally instead of returning a row and column and letting the game module handle that. This makes it easier to upgrade the CPU random attack logic, as it captures whether or not the attack is a hit, and stores that data for future attacks. randomAttack now returns the value received from GameBoard.recieveAttack.